% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function from Lab 9}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{vector representation of x}

\item{fun}{default value "mean"}

\item{alpha}{default value 0.05}

\item{cx}{default value 1.5}

\item{...}{...}
}
\value{
Will make a number of plots, point and interval estimates and useful
output in the form of a list.
}
\description{
Bootstrap Function from Lab 9
}
\examples{
\dontrun{set.seed(39); sam = rnorm(25, mean = 25, sd = 10)
windows(); myboot2(10000, x = sam, fun = "mean", alpha = 0.05, xlab = "mean",
col = "Purple", cx = 1.5)}
}
