% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem function from Lab 8}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{size}

\item{iter}{iterations}
}
\value{
plot
}
\description{
Central Limit Theorem function from Lab 8
}
\examples{
\dontrun{w = myclt(n = 10, iter = 10000)}
}
