% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Central Limit Theorem Binomial function from Lab 8}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{size}

\item{iter}{iterations}

\item{p}{probability}

\item{...}{...}
}
\value{
plot
}
\description{
Central Limit Theorem Binomial function from Lab 8
}
\examples{
\dontrun{mycltb(n = 5, iter = 10000, p = 0.5)}
}
