% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem Poisson function from Lab 8}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{size}

\item{iter}{iterations}

\item{lambda}{lambda for the poisson distribution}

\item{...}{...}
}
\value{
plot
}
\description{
Central Limit Theorem Poisson function from Lab 8
}
\examples{
\dontrun{mycltp(n = 10, iter = 10000)}
}
