% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Function Uniform from Lab 8}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{size}

\item{iter}{iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
plot
}
\description{
Central Limit Function Uniform from Lab 8
}
\examples{
\dontrun{mycltu(n = 20, iter = 100000)}
}
